refactor(core): split stand alone mode function for more clarity after review
